refactor: moved pokemon card component inside results component
