+++ src/app/components/results/pokemon-card/pokemon-card.component.scss
