+++ src/app/components/results/pokemon-card/pokemon-card.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { DamageDetails, Type } from '../../types/types';
+import { DamageDetails, Type } from '../../../types/types';
 
 @Component({
   selector: 'app-pokemon-card',
